Features/search engine

+++ fullstack_project/backend/app/routers/ratings.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from typing import List
 from app.core.security import verify_access_token
 from app.schemas.rating import Rating, RatingCreate, RatingUpdate
@@ -12,8 +12,8 @@ def get_Ratings():
 
 #simple post the payload (is the body of the request)
 @router.post("", response_model=Rating, status_code=201)
-def post_rating(payload: RatingCreate):
-    return create_rating(payload)
+def post_rating(payload: RatingCreate, token_data : dict = Depends(verify_access_token)):
+    return create_rating(payload, token_data['user'])
 
 @router.get("/isbn/{isbn}", response_model=List)
 def get_rating_by_isbn(isbn: str):
@@ -24,13 +24,19 @@ def get_rating_by_id(id: str):
     return get_rating_by_id(id)
 
 ## We use put here because we are not creating an entirely new item, ie. we keep id the same
-@router.put("/{isbn}/{id}", response_model=Rating)
-def put_rating(isbn: str, id: str, payload: RatingUpdate):
-    return update_rating(isbn, id, payload)
-
+@router.put("/{isbn}", response_model=Rating)
+def put_rating(isbn: str, id: str, payload: RatingUpdate, token_data : dict = Depends(verify_access_token)):
+    return update_rating(isbn, token_data['user'], payload)
 
 ## we put the status there becuase in a delete, we wont have a return so it indicates it happened succesfully
 @router.delete("/{isbn}/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def remove_rating(isbn: str, id: str):
+def remove_rating_admin(isbn : str, id: str, token_data : dict = Depends(verify_access_token)):
+    if not token_data["is_admin"] :
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin access required")
     delete_rating(isbn, id)
+    return None
+
+@router.delete("/{isbn}", status_code=status.HTTP_204_NO_CONTENT)
+def remove_rating(isbn : str, token_data : dict = Depends(verify_access_token)):
+    delete_rating(isbn, token_data['user'])
     return None
